Change VirtualCalls and FastHierarchy to work with SootMethodRef instead of SootMethod in it's interface. This prevents calls to SootMethodRef.resolve when receiving the SootMethod from an InvokeInstruction.
SootMethodRef.resolve leads to the instrumentation of phantom SootMethods if the Method could not be dispatched correctly due to missing depenendencies on the classpath. This eventually leads to these phantom Methods being targets in the callgraph. This is not a bad thing on its own, but in the current develop branch these methods would just be omitted from the callgraph, so we should try staying consistent with that if some client rely on the behavior.